Add migrator for kube shell image and add it as a default in the db
