+++ api/bolt/migrator/migrate_dbversion33.go
@@ -1,32 +1,21 @@
 package migrator
 
-import (
-	portainer "github.com/portainer/portainer/api"
-)
+import portainer "github.com/portainer/portainer/api"
 
-func (m *Migrator) migrateDBVersionTo33() error {
-	err := migrateStackEntryPoint(m.stackService)
-	if err != nil {
+func (m *Migrator) migrateDBVersionToDB33() error {
+	if err := m.migrateSettingsToDB33(); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func migrateStackEntryPoint(stackService portainer.StackService) error {
-	stacks, err := stackService.Stacks()
+func (m *Migrator) migrateSettingsToDB33() error {
+	settings, err := m.settingsService.Settings()
 	if err != nil {
 		return err
 	}
-	for i := range stacks {
-		stack := &stacks[i]
-		if stack.GitConfig == nil {
-			continue
-		}
-		stack.GitConfig.ConfigFilePath = stack.EntryPoint
-		if err := stackService.UpdateStack(stack.ID, stack); err != nil {
-			return err
-		}
-	}
-	return nil
+
+	settings.KubectlShellImage = portainer.DefaultKubectlShellImage
+	return m.settingsService.UpdateSettings(settings)
 }
